locked down non admins to only be able to do no get methods in only their dealership

+++ DealershipApi.Services/Services/VehicleService.cs
@@ -1,5 +1,6 @@
 ﻿using DealershipApi.Data.DataModels;
 using DealershipApi.Models.DisplayModels.Vehicle;
+using Microsoft.AspNet.Identity.EntityFramework;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -21,11 +22,15 @@ namespace DealershipApi.Services.Services
         {
             using (var ctx = new ApplicationDbContext())
             {
-                var loggedInUser =
-                        ctx
-                            .Users
-                            .Single(u => u.Id == _userId);
-                
+
+                var loggedInUser = (from u in ctx.Users
+                                    let query = (from ur in ctx.Set<IdentityUserRole>()
+                                                 where ur.UserId.Equals(u.Id)
+                                                 join r in ctx.Roles on ur.RoleId equals r.Id
+                                                 select r.Name)
+                                    where u.Id == _userId
+                                    select new { u.Id, u.DealershipId, listOfRoles = query.ToList() }).Single();
+
 
                 var entity =
                     new Vehicle()
@@ -41,8 +46,11 @@ namespace DealershipApi.Services.Services
                         VehicleCondition = model.VehicleCondition,
                         InStock = true,
                     };
-
-                ctx.Vehicles.Add(entity);
+                if (loggedInUser.DealershipId == entity.DealershipId || loggedInUser.listOfRoles.Contains("Admin"))
+                {
+                    ctx.Vehicles.Add(entity);
+                }
+                
                 return ctx.SaveChanges() == 1;
             }
         }
@@ -71,24 +79,27 @@ namespace DealershipApi.Services.Services
         {
             using (var ctx = new ApplicationDbContext())
             {
+              
+               
                 var entity =
                     ctx
                         .Vehicles
                         .Single(v => v.Id == id);
-
-                return
-                    new VehicleDetail
-                    {
-                        Id = entity.Id,
-                        Make = entity.Make,
-                        Model = entity.Model,
-                        ModelYear = entity.ModelYear,
-                        Color = entity.Color,
-                        SalesPrice = entity.InvoicePrice,
-                        DealershipId = entity.DealershipId,
-                        VehicleCondition = entity.VehicleCondition,
-                        Mileage = entity.Mileage,
-                    };
+                
+                    return
+                        new VehicleDetail
+                        {
+                            Id = entity.Id,
+                            Make = entity.Make,
+                            Model = entity.Model,
+                            ModelYear = entity.ModelYear,
+                            Color = entity.Color,
+                            SalesPrice = entity.InvoicePrice,
+                            DealershipId = entity.DealershipId,
+                            VehicleCondition = entity.VehicleCondition,
+                            Mileage = entity.Mileage,
+                        };
+                return null;
             }
         }
 
@@ -96,20 +107,30 @@ namespace DealershipApi.Services.Services
         {
             using (var ctx = new ApplicationDbContext())
             {
+                var loggedInUser = (from u in ctx.Users
+                                    let query = (from ur in ctx.Set<IdentityUserRole>()
+                                                 where ur.UserId.Equals(u.Id)
+                                                 join r in ctx.Roles on ur.RoleId equals r.Id
+                                                 select r.Name)
+                                    where u.Id == _userId
+                                    select new { u.Id, u.DealershipId, listOfRoles = query.ToList() }).Single();
                 var entity =
                     ctx
                         .Vehicles
                         .Single(v => v.Id == model.Id);
-
-                entity.Id = model.Id;
-                entity.Make = model.Make;
-                entity.Model = model.Model;
-                entity.ModelYear = model.ModelYear;
-                entity.Color = model.Color;
-                entity.InvoicePrice = model.SalesPrice;
-                entity.DealershipId = model.DealershipId;
-                entity.VehicleCondition = model.VehicleCondition;
-                entity.Mileage = model.Mileage;
+                if (loggedInUser.DealershipId == entity.DealershipId || loggedInUser.listOfRoles.Contains("Admin"))
+                {
+                    entity.Id = model.Id;
+                    entity.Make = model.Make;
+                    entity.Model = model.Model;
+                    entity.ModelYear = model.ModelYear;
+                    entity.Color = model.Color;
+                    entity.InvoicePrice = model.SalesPrice;
+                    entity.DealershipId = model.DealershipId;
+                    entity.VehicleCondition = model.VehicleCondition;
+                    entity.Mileage = model.Mileage;
+                }
+              
 
                 return ctx.SaveChanges() == 1;
             }
@@ -119,11 +140,22 @@ namespace DealershipApi.Services.Services
         {
             using (var ctx = new ApplicationDbContext())
             {
+                var loggedInUser = (from u in ctx.Users
+                                    let query = (from ur in ctx.Set<IdentityUserRole>()
+                                                 where ur.UserId.Equals(u.Id)
+                                                 join r in ctx.Roles on ur.RoleId equals r.Id
+                                                 select r.Name)
+                                    where u.Id == _userId
+                                    select new { u.Id, u.DealershipId, listOfRoles = query.ToList() }).Single();
                 var entity =
                     ctx
                         .Vehicles
                         .Single(v => v.Id == id);
-                ctx.Vehicles.Remove(entity);
+                if(loggedInUser.DealershipId == entity.DealershipId || loggedInUser.listOfRoles.Contains("Admin"))
+                {
+                    ctx.Vehicles.Remove(entity);
+                }
+                
                 return ctx.SaveChanges() == 1;
             }
         }
